pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "python-deriv-api>=0.1.6",
]